refactor: fix eslint error for useAuth and AuthProvider exports
